+++ daffodil/__init__.py
@@ -1,7 +1,7 @@
 from .predicate import DictionaryPredicateDelegate
 from .hstore_predicate import HStoreQueryDelegate
 from .pretty_print import PrettyPrintDelegate
-from .parser import Daffodil
+from .peg_parser import Daffodil
 
 # for backwards compatibility. probably unnecessary.
 from parsimonious.grammar import Grammar
